CXP-1102: move error management command
