+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/DependencyInjection/AkeneoConnectivityConnectionExtension.php
@@ -36,6 +36,8 @@ class AkeneoConnectivityConnectionExtension extends Extension
         $loader->load('Audit/install.yml');
         $loader->load('Audit/queries.yml');
 
+        $loader->load('ErrorManagement/commands.yml');
+
         $loader->load('Marketplace/controllers.yml');
         $loader->load('Marketplace/handlers.yml');
         $loader->load('Marketplace/install.yml');
