+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/cli.yml
@@ -5,10 +5,3 @@ services:
             - '@akeneo_connectivity.connection.application.handler.create_connection'
         tags:
             - { name: 'console.command' }
-
-    Akeneo\Connectivity\Connection\Infrastructure\Cli\PurgeConnectionErrorsCommand:
-        arguments:
-            - '@akeneo_connectivity_connection.persistence.query.select_all_auditable_connection_code'
-            - '@akeneo_connectivity_connection.persistence.query.purge_connection_errors'
-        tags:
-            - {name: 'console.command'}
